Removed DockerManager file
