+++ e2e/helpers/environment/EnvironmentManager.ts
@@ -1,12 +1,31 @@
 import {ContainerState} from './ContainerState';
-import {DockerManager, GhostContainerConfig} from './DockerManager';
 import Docker from 'dockerode';
+import type {Container} from 'dockerode';
 import logging from '@tryghost/logging';
 import baseDebug from '@tryghost/debug';
 import path from 'path';
 import {execSync} from 'child_process';
 
 const debug = baseDebug('e2e:EnvironmentManager');
+
+export interface GhostContainerConfig {
+    networkId: string;
+    networkAlias: string;
+    database: string;
+    mysqlHost: string;
+    mysqlPort: string;
+    mysqlUser: string;
+    mysqlPassword: string;
+    exposedPort: number;
+    siteUuid: string;
+    workingDir?: string;
+    command?: string[];
+    tinybird: {
+        workspaceId: string;
+        adminToken: string;
+        trackerToken: string;
+    };
+}
 
 export interface GhostInstance {
     containerId: string;
@@ -18,14 +37,12 @@ export interface GhostInstance {
 
 export class EnvironmentManager {
     private containerState: ContainerState;
-    private dockerManager: DockerManager;
     private composeFilePath = path.resolve(__dirname, '../../compose.e2e.yml');
     private composeProjectName = 'ghost-e2e';
     private docker: Docker;
 
     constructor() {
         this.containerState = new ContainerState();
-        this.dockerManager = new DockerManager();
         this.docker = new Docker();
     }
 
@@ -35,13 +52,13 @@ export class EnvironmentManager {
      * @returns The Docker container entity
      * @throws Error if no container or multiple containers are found for the service
      */
-    public async getContainerForService(service: string): Promise<Docker.Container> {
+    public async getContainerForService(service: string): Promise<Container> {
         debug('getContainerForService called for service:', service);
-        const containers = await this.docker.listContainers({all: true, filters: {label: [
-            `com.docker.compose.project=${this.composeProjectName}`,
-            `com.docker.compose.service=${service}`
-        ]}});
-        debug('getContainerForService found containers:', containers.map(c => c.Id));
+        const containers = await this.docker.listContainers({all: true,
+            filters: {label: [
+                `com.docker.compose.project=${this.composeProjectName}`,
+                `com.docker.compose.service=${service}`
+            ]}});
         if (containers.length === 0) {
             throw new Error(`No container found for service: ${service}`);
         }
@@ -155,26 +172,25 @@ export class EnvironmentManager {
      * Setup a Ghost instance for a specific test
      * This method is designed to be called from test fixtures
      */
-    public async setupGhostInstance(workerId: number, testId: string): Promise<GhostInstance> {
+    public async setupGhostInstance(workerId: number): Promise<GhostInstance> {
         try {
-            debug('Setting up Ghost instance for worker %d, test %s', workerId, testId);
+            debug('Setting up Ghost instance for worker %d, test %s', workerId);
             const network = await this.getNetwork();
-            const networkAlias = ContainerState.generateNetworkAlias(workerId, testId);
+            const siteUuid = crypto.randomUUID();
+            const instanceId = `ghost_${siteUuid}`;
             const port = ContainerState.generateUniquePort(workerId);
-            const siteUuid = ContainerState.generateSiteUuid();
-            const database = `ghost_${siteUuid}`;
             const tinybirdState = this.containerState.loadTinybirdState();
 
-            debug('Generated test-specific identifiers:', {siteUuid, database, networkAlias, port});
+            debug('Generated test-specific identifiers:', {siteUuid, instanceId, port});
 
             // Create and restore database
-            await this.setupTestDatabase(database, siteUuid);
+            await this.setupTestDatabase(instanceId, siteUuid);
 
             // Create Ghost container
             const ghostConfig: GhostContainerConfig = {
                 networkId: network.id,
-                networkAlias: networkAlias,
-                database: database,
+                networkAlias: instanceId,
+                database: instanceId,
                 mysqlHost: 'mysql',
                 mysqlPort: '3306',
                 mysqlUser: 'root',
@@ -190,11 +206,11 @@ export class EnvironmentManager {
                 }
             };
 
-            const containerId = await this.dockerManager.createGhostContainer(ghostConfig);
+            const containerId = await this.createGhostContainer(ghostConfig);
 
             const ghostInstance: GhostInstance = {
                 containerId,
-                database,
+                database: instanceId,
                 port,
                 baseUrl: `http://localhost:${port}`,
                 siteUuid
@@ -219,7 +235,7 @@ export class EnvironmentManager {
             debug('Tearing down Ghost instance:', ghostInstance.containerId);
 
             // Stop and remove the Ghost container
-            await this.dockerManager.removeContainer(ghostInstance.containerId);
+            await this.removeContainer(ghostInstance.containerId);
 
             // Clean up the database
             await this.cleanupTestDatabase(ghostInstance.database);
@@ -291,7 +307,7 @@ export class EnvironmentManager {
      * @returns The command output
      * @throws Error if the command fails
      */
-    private async execInContainer(container: Docker.Container, command: string): Promise<string> {
+    private async execInContainer(container: Container, command: string): Promise<string> {
         // Wrap command with exit code check
         const wrappedCommand = `${command}; echo "__EXIT_CODE__=$?"`;
 
@@ -328,6 +344,147 @@ export class EnvironmentManager {
         // If no exit code found, return output as-is but log a warning
         debug('Warning: No exit code found in command output');
         return output;
+    }
+
+    /**
+     * Create and start a Ghost container
+     */
+    private async createGhostContainer(config: GhostContainerConfig): Promise<string> {
+        try {
+            const environment = {
+                server__host: '0.0.0.0',
+                server__port: '2368',
+                database__client: 'mysql2',
+                database__connection__host: config.mysqlHost,
+                database__connection__port: config.mysqlPort,
+                database__connection__user: config.mysqlUser,
+                database__connection__password: config.mysqlPassword,
+                database__connection__database: config.database,
+                NODE_ENV: 'development',
+                // Tinybird configuration
+                TB_HOST: 'http://tinybird-local:7181',
+                TB_LOCAL_HOST: 'tinybird-local',
+                tinybird__stats__endpoint: 'http://tinybird-local:7181',
+                tinybird__stats__endpointBrowser: 'http://localhost:7181',
+                tinybird__tracker__endpoint: 'http://localhost/.ghost/analytics/api/v1/page_hit',
+                tinybird__workspaceId: config.tinybird.workspaceId,
+                tinybird__adminToken: config.tinybird.adminToken
+            };
+
+            const containerConfig = {
+                Image: process.env.GHOST_IMAGE_TAG || 'ghost-monorepo',
+                Env: Object.entries(environment).map(([key, value]) => `${key}=${value}`),
+                NetworkingConfig: {
+                    EndpointsConfig: {
+                        [config.networkId]: {
+                            Aliases: [config.networkAlias]
+                        }
+                    }
+                },
+                ExposedPorts: {
+                    '2368/tcp': {}
+                },
+                HostConfig: {
+                    PortBindings: {
+                        '2368/tcp': [{HostPort: config.exposedPort.toString()}]
+                    }
+                },
+                Labels: {
+                    'com.docker.compose.project': 'ghost-e2e',
+                    'com.docker.compose.service': `ghost-${config.siteUuid}`,
+                    'tryghost/e2e': 'ghost'
+                },
+                WorkingDir: config.workingDir || '/home/ghost/ghost/core',
+                Cmd: config.command || ['yarn', 'dev'],
+                AttachStdout: true,
+                AttachStderr: true
+            };
+
+            debug('Creating Ghost container with config:', {
+                networkId: config.networkId,
+                alias: config.networkAlias,
+                database: config.database,
+                port: config.exposedPort
+            });
+
+            debug('Ghost environment variables:', JSON.stringify(environment, null, 2));
+
+            debug('Full Docker container config:', JSON.stringify(containerConfig, null, 2));
+
+            const container = await this.docker.createContainer(containerConfig);
+            await container.start();
+
+            // Inspect the container to see actual port mappings
+            const containerInfo = await container.inspect();
+            debug('Container port mappings:', containerInfo.NetworkSettings.Ports);
+
+            // Wait for Ghost to be ready by checking HTTP health
+            await this.waitForGhostReady(config.exposedPort);
+
+            debug('Ghost container started successfully:', container.id);
+            return container.id;
+        } catch (error) {
+            logging.error('Failed to create Ghost container:', error);
+            throw new Error(`Failed to create Ghost container: ${error}`);
+        }
+    }
+
+    /**
+     * Wait for Ghost to be ready by checking HTTP health
+     */
+    private async waitForGhostReady(port: number, timeoutMs: number = 60000): Promise<void> {
+        const startTime = Date.now();
+        const healthUrl = `http://localhost:${port}/ghost/api/admin/site/`;
+
+        while (Date.now() - startTime < timeoutMs) {
+            try {
+                // Simple HTTP check to see if Ghost API responds
+                const response = await fetch(healthUrl, {
+                    method: 'GET',
+                    signal: AbortSignal.timeout(5000) // 5 second timeout per request
+                });
+
+                // If we get any response (even 401/404), Ghost is ready
+                if (response.status < 500) {
+                    debug('Ghost is ready, responded with status:', response.status);
+                    return;
+                }
+
+                debug('Ghost not ready yet, status:', response.status);
+            } catch (error) {
+                debug('Ghost health check failed, retrying...', error instanceof Error ? error.message : String(error));
+            }
+
+            // Wait 200ms before next check
+            await new Promise<void>((resolve) => {
+                setTimeout(resolve, 200);
+            });
+        }
+
+        throw new Error(`Timeout waiting for Ghost to start on port ${port}`);
+    }
+
+    /**
+     * Stop and remove a container
+     */
+    private async removeContainer(containerId: string): Promise<void> {
+        try {
+            const container = this.docker.getContainer(containerId);
+
+            // Stop the container (with force if needed)
+            try {
+                await container.stop({t: 10}); // 10 second timeout
+            } catch (error) {
+                debug('Container already stopped or stop failed, forcing removal:', containerId);
+            }
+
+            // Remove the container
+            await container.remove({force: true});
+            debug('Container removed:', containerId);
+        } catch (error) {
+            debug('Failed to remove container:', error);
+            // Don't throw - container might already be removed
+        }
     }
 }
