+++ e2e/package.json
@@ -29,7 +29,8 @@
     "ts-node": "10.9.2",
     "typescript": "5.8.3",
     "@tryghost/logging": "2.4.23",
-    "@tryghost/debug": "0.1.35"
+    "@tryghost/debug": "0.1.35",
+    "@types/dockerode": "3.3.43"
   },
   "dependencies": {}
 }
